Moved 'WriteVdfV1()' into separate file.
