+++ shortcuts/shortcut.go
@@ -1,7 +1,6 @@
 package shortcuts
 
 import (
-	"io"
 	"strings"
 
 	"github.com/stephen-fox/steamutil/vdf"
@@ -146,36 +145,4 @@ func appendDoubleQuotesIfNeeded(s string) string {
 	}
 
 	return s
-}
-
-func WriteVdfV1(shortcuts []Shortcut, w io.Writer) error {
-	config := vdf.Config{
-		Name:    header,
-		Version: vdf.V1,
-	}
-
-	c, err := vdf.NewConstructor(config)
-	if err != nil {
-		return err
-	}
-
-	var objects []vdf.Object
-
-	for _, s := range shortcuts {
-		objects = append(objects, s.object())
-	}
-
-	v := c.Build(objects)
-
-	s, err := v.String()
-	if err != nil {
-		return err
-	}
-
-	_, err = w.Write([]byte(s))
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
